Enhance Chart function
Forecast data  in array of 40 items doesnt always start at the same time, sometimes it starts at 6:00, the other time at 9:00 or 3:00. this leads to the incorrect data retrieve. Setting fix time point to display forecast which is 6 items later than the 1st item is sometimes not efficient as this 2nd item still belongs to the same date as the 1st.